# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HustleCoin Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramInitData {
  telegramId: string;
  username: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  referredBy?: string | null;
  /** @nullable */
  initData?: string | null;
}

export interface UserProfile {
  id: number;
  telegramId: string;
  username: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  balance: number;
  level: number;
  streak: number;
  totalMines: number;
  /** @nullable */
  lastMine?: string | null;
  canMine?: boolean;
  /** @nullable */
  mineCountdown?: number | null;
  referralCount: number;
  achievementCount: number;
  joinDate: string;
  /** @nullable */
  rank: number | null;
  badges?: string[];
}

export interface MineRequest {
  telegramId: string;
}

export interface MineResult {
  success: boolean;
  hpEarned: number;
  bonusHp: number;
  newBalance: number;
  streak: number;
  message: string;
  /** @nullable */
  nextMineAt?: string | null;
}

export interface MineLog {
  id: number;
  telegramId: string;
  hpEarned: number;
  bonusHp?: number;
  streak: number;
  minedAt: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  link?: string | null;
  status: TaskStatus;
  createdAt?: string;
}

export type TaskWithStatusStatus = typeof TaskWithStatusStatus[keyof typeof TaskWithStatusStatus];


export const TaskWithStatusStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface TaskWithStatus {
  id: number;
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  link?: string | null;
  status: TaskWithStatusStatus;
  completed: boolean;
  approved: boolean;
  /** @nullable */
  completedAt?: string | null;
}

export interface TaskInput {
  title: string;
  description: string;
  reward: number;
  /** @nullable */
  link?: string | null;
  adminTelegramId: string;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface TaskUpdate {
  title?: string;
  description?: string;
  reward?: number;
  /** @nullable */
  link?: string | null;
  status?: TaskUpdateStatus;
  adminTelegramId?: string;
}

export interface TaskCompleteRequest {
  telegramId: string;
}

export interface TaskCompletion {
  id: number;
  taskId: number;
  telegramId: string;
  approved: boolean;
  completedAt: string;
}

export interface ApproveTaskRequest {
  telegramId: string;
  adminTelegramId: string;
}

export type QuestWithProgressQuestType = typeof QuestWithProgressQuestType[keyof typeof QuestWithProgressQuestType];


export const QuestWithProgressQuestType = {
  mine: 'mine',
  complete_task: 'complete_task',
  invite_friend: 'invite_friend',
} as const;

export interface QuestWithProgress {
  id: number;
  title: string;
  description: string;
  reward: number;
  questType: QuestWithProgressQuestType;
  completed: boolean;
  progress: number;
  target: number;
}

export interface AchievementWithStatus {
  id: number;
  title: string;
  description: string;
  icon: string;
  unlocked: boolean;
  /** @nullable */
  unlockedAt?: string | null;
}

export interface ReferralEntry {
  telegramId: string;
  username: string;
  firstName: string;
  joinDate: string;
  hpEarned: number;
}

export interface ReferralStats {
  telegramId: string;
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  totalEarned: number;
  referrals: ReferralEntry[];
}

export interface LeaderboardEntry {
  rank: number;
  telegramId: string;
  username: string;
  firstName: string;
  value: number;
  /** @nullable */
  badge?: string | null;
}

export interface FeedbackInput {
  telegramId: string;
  message: string;
}

export interface Feedback {
  id: number;
  telegramId: string;
  message: string;
  /** @nullable */
  adminReply?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  user: UserProfile;
  questsCompleted: number;
  questsTotal: number;
  activeTasksCount: number;
  recentAchievements: AchievementWithStatus[];
}

export interface AdminStats {
  totalUsers: number;
  newUsersToday: number;
  newUsersThisWeek: number;
  activeUsersToday: number;
  activeUsersThisWeek: number;
  averageHp: number;
  totalReferrals: number;
  totalAchievements: number;
  totalMines: number;
}

export interface AdminUser {
  id: number;
  telegramId: string;
  username: string;
  firstName: string;
  balance: number;
  level: number;
  streak: number;
  joinDate: string;
  totalMines: number;
  referralCount: number;
}

export interface GrantHpRequest {
  adminTelegramId: string;
  targetTelegramId: string;
  amount: number;
  reason: string;
}

export interface BroadcastInput {
  adminTelegramId: string;
  message: string;
}

export interface BroadcastResult {
  success: boolean;
  message: string;
}

export type GetProfileParams = {
telegramId: string;
};

export type GetMineHistoryParams = {
telegramId: string;
limit?: number;
};

export type GetTasksParams = {
telegramId: string;
};

export type GetQuestsParams = {
telegramId: string;
};

export type GetAchievementsParams = {
telegramId: string;
};

export type GetReferralsParams = {
telegramId: string;
};

export type GetLeaderboardParams = {
type?: GetLeaderboardType;
limit?: number;
};

export type GetLeaderboardType = typeof GetLeaderboardType[keyof typeof GetLeaderboardType];


export const GetLeaderboardType = {
  hp: 'hp',
  referrals: 'referrals',
  streak: 'streak',
  mining: 'mining',
} as const;

export type GetDashboardParams = {
telegramId: string;
};

export type GetAdminStatsParams = {
telegramId: string;
};

export type GetAdminUsersParams = {
telegramId: string;
limit?: number;
offset?: number;
};

export type GetAdminFeedbackParams = {
telegramId: string;
};