# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleCoin Telegram Mini App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mine:
    post:
      operationId: mineHp
      tags: [mining]
      summary: Mine HP coins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MineRequest"
      responses:
        "200":
          description: Mining result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MineResult"
        "400":
          description: Cannot mine yet (cooldown active)
  /mine/history:
    get:
      operationId: getMineHistory
      tags: [mining]
      summary: Get mining history for a user
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Mining history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MineLog"
...
  /feedback:
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit user feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin growth analytics
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "403":
          description: Forbidden
  /admin/grant-hp:
    post:
      operationId: grantHp
      tags: [admin]
      summary: Grant or remove HP from a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantHpRequest"
      responses:
        "200":
          description: HP updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
  /admin/broadcast:
    post:
      operationId: broadcastMessage
      tags: [admin]
      summary: Broadcast announcement to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
        "403":
          description: Forbidden
  /admin/feedback:
    get:
      operationId: getAdminFeedback
      tags: [admin]
      summary: Get all feedback (admin)
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: All feedback
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, telegramId, username, firstName, balance, level, streak, totalMines, referralCount, achievementCount, joinDate, rank]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: ["string", "null"]
        balance:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        totalMines:
          type: integer
        lastMine:
          type: ["string", "null"]
        canMine:
          type: boolean
        mineCountdown:
          type: ["integer", "null"]
        referralCount:
          type: integer
        achievementCount:
          type: integer
        joinDate:
          type: string
        rank:
          type: ["integer", "null"]
        badges:
          type: array
          items:
            type: string
    MineRequest:
      type: object
      required: [telegramId]
      properties:
        telegramId:
          type: string
    MineResult:
      type: object
      required: [success, hpEarned, newBalance, streak, bonusHp, message]
      properties:
        success:
          type: boolean
        hpEarned:
          type: integer
        bonusHp:
          type: integer
        newBalance:
          type: integer
        streak:
          type: integer
        message:
          type: string
        nextMineAt:
          type: ["string", "null"]
    MineLog:
      type: object
      required: [id, telegramId, hpEarned, minedAt, streak]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        hpEarned:
          type: integer
        bonusHp:
          type: integer
        streak:
          type: integer
        minedAt:
          type: string
...
    FeedbackInput:
      type: object
      required: [telegramId, message]
      properties:
        telegramId:
          type: string
        message:
          type: string
    Feedback:
      type: object
      required: [id, telegramId, message, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        message:
          type: string
        adminReply:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AdminStats:
      type: object
      required: [totalUsers, newUsersToday, newUsersThisWeek, activeUsersToday, activeUsersThisWeek, averageHp, totalReferrals, totalAchievements, totalMines]
      properties:
        totalUsers:
          type: integer
        newUsersToday:
          type: integer
        newUsersThisWeek:
          type: integer
        activeUsersToday:
          type: integer
        activeUsersThisWeek:
          type: integer
        averageHp:
          type: integer
        totalReferrals:
          type: integer
        totalAchievements:
          type: integer
        totalMines:
          type: integer
    AdminUser:
      type: object
      required: [id, telegramId, username, firstName, balance, level, streak, joinDate, totalMines, referralCount]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: string
        firstName:
          type: string
        balance:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        joinDate:
          type: string
        totalMines:
          type: integer
        referralCount:
          type: integer
    GrantHpRequest:
      type: object
      required: [adminTelegramId, targetTelegramId, amount, reason]
      properties:
        adminTelegramId:
          type: string
        targetTelegramId:
          type: string
        amount:
          type: integer
        reason:
          type: string
    BroadcastInput:
      type: object
      required: [adminTelegramId, message]
      properties:
        adminTelegramId:
          type: string
        message:
          type: string
    BroadcastResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string